project/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==24.1.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-redis==5.4.0
django-rosetta==0.10.0
django-timezone-field==7.0
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
packaging==24.0
pillow==10.2.0
polib==1.2.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.31.0
rpds-py==0.19.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.4.4
transliterate==1.10.2
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
